+++ include/mstd/types.hpp
@@ -1,0 +1,6 @@
+#ifndef __MSTD__TYPES_HPP__
+#define __MSTD__TYPES_HPP__
+
+#include "types/strong_type.hpp"   // IWYU pragma: export
+
+#endif   // __MSTD__TYPES_HPP__